\name{badgertub-package}
\alias{badgertub-package}
\alias{badgertub}
\docType{package}
\title{
\packageTitle{badgertub}
}
\description{
\packageDescription{badgertub}
}
\details{

  This package is a companion package to the paper of Calenge et al. (in
  prep.), containing the data and code used to fit the models and carry
  out the simulations carried out in this package.  The vignette (named
  "badgertub.pdf") describes how to use this code, and also serves as
  the appendix of this paper.
  
}
\author{
\packageAuthor{badgertub}

Maintainer: \packageMaintainer{badgertub}
}
\references{
  
  Calenge C., Desvaux S., Payne A., Reveillaud E. and Richomme C. (in
  prep.) Assessing the bovine tuberculosis dynamics in three French badger
  populations with a participatory monitoring programme.

}
\keyword{ package }
\examples{

\dontrun{
vignette("badgertub")
}
}
